// packages/test/generatedCode/identity/b2xUserFlows/item/languages/item/userFlowLanguageConfigurationItemRequestBuilderGetRequestConfiguration.ts
import {UserFlowLanguageConfigurationItemRequestBuilderGetQueryParameters} from './userFlowLanguageConfigurationItemRequestBuilderGetQueryParameters';
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface UserFlowLanguageConfigurationItemRequestBuilderGetRequestConfiguration {
    /** Request headers */
    headers?: Record<string, string>;
    /** Request options */
    options?: RequestOption[];
    /** Request query parameters */
    queryParameters?: UserFlowLanguageConfigurationItemRequestBuilderGetQueryParameters;
}